src/main/java/com/github/mpi/time_registration/domain/WorkLogEntryFactory.java
package com.github.mpi.time_registration.domain;

import com.github.mpi.time_registration.domain.time.Day;

public class WorkLogEntryFactory {

    private final EntryIDSequence sequence;
    private final EmployeeContext employeeContext;

    public WorkLogEntryFactory(EntryIDSequence sequence, EmployeeContext employeeContext) {
        this.sequence = sequence;
        this.employeeContext = employeeContext;
    }

    public WorkLogEntry newEntry(String workload, String projectName) {
        return new WorkLogEntry(sequence.nextID(), Workload.of(workload), new ProjectName(projectName), employeeContext.employeeID(), null);
        
    }

    public WorkLogEntry newEntry(String workload, String projectName, String day) {
        return new WorkLogEntry(sequence.nextID(), Workload.of(workload), new ProjectName(projectName), employeeContext.employeeID(), Day.of(day));
    }
    
}
